refactor(): * split out measurement (area, volume etc) functions from CSG & CAG * split out csg projection function from CSG * various cleanups
